+++ src/rr_core/controller_manager.rs
@@ -1,24 +1,25 @@
 extern crate hidapi;
-use hidapi::{HidApi, HidDevice, DeviceInfo};
+use hidapi::{HidApi, DeviceInfo};
 
-use crate::rr_core::interface::{RRMessage,AppState,ControllerMessage,RGB ,ControllerConnectionType, DualShock4, Dpad, JoyStick, Buttons};
+use super::controller_driver::{interface::{Controller, ControllerConnectionType, RGB}, self};
+use crate::rr_core::interface::{RRMessage,AppState,ControllerMessage};
 use crate::rr_core::{utils, thread_connection};
 
 use iced_aw::TabLabel;
 
-pub struct DualShock4DriverManager
+pub struct ControllerManager
 {
-    pub first_connector:thread_connection::AsyncThreadConnector<DualShock4>,
-    pub connectors:Vec<thread_connection::ThreadConnector<DualShock4>>,
+    pub first_connector:thread_connection::AsyncThreadConnector<Controller>,
+    pub connectors:Vec<thread_connection::ThreadConnector<Controller>>,
     pub controller_num:usize,
     pub device_list:Vec<DeviceInfo>,
     api:HidApi,
-    pub get_value:Vec<DualShock4>,
+    pub get_value:Vec<Controller>,
     pub state:AppState,
     green_flag:bool
 }
 
-impl DualShock4DriverManager {
+impl ControllerManager {
     fn title(&self)->String
     {
         String::from("コントローラー設定")
@@ -97,18 +98,18 @@ impl DualShock4DriverManager {
                     {   
                         self.spawn_driver(ControllerConnectionType::SERIAL);
                         self.controller_num += 1;
-                        self.get_value.push(DualShock4::new());
+                        self.get_value.push(Controller::new());
                         self.device_list.remove(0);
 
                         for i in 0..self.device_list.len() {
-                            let new_conn = thread_connection::ThreadConnector::<DualShock4>::new();
+                            let new_conn = thread_connection::ThreadConnector::<Controller>::new();
                             self.connectors.push(new_conn);
 
                             self.add_driver(ControllerConnectionType::SERIAL, self.connectors.get(i+1).unwrap().publisher.clone());
                             self.device_list.remove(0);
 
                             self.controller_num += 1;
-                            self.get_value.push(DualShock4::new());
+                            self.get_value.push(Controller::new());
 
                             self.state = AppState::OK;
                         }
@@ -123,17 +124,17 @@ impl DualShock4DriverManager {
     }
 }
 
-impl DualShock4DriverManager {
-    pub fn new()->DualShock4DriverManager
+impl ControllerManager {
+    pub fn new()->ControllerManager
     {
-        let mut ds4_conn_vec = Vec::<thread_connection::ThreadConnector<DualShock4>>::new();
-        let ds4_conn = thread_connection::AsyncThreadConnector::<DualShock4>::new();
-        let sync_conn = thread_connection::ThreadConnector::<DualShock4>::new();
+        let mut ds4_conn_vec = Vec::<thread_connection::ThreadConnector<Controller>>::new();
+        let ds4_conn = thread_connection::AsyncThreadConnector::<Controller>::new();
+        let sync_conn = thread_connection::ThreadConnector::<Controller>::new();
         ds4_conn_vec.push(sync_conn);
-        let mut in_v = Vec::<DualShock4>::new();
-        let ds4 = DualShock4::new();
+        let mut in_v = Vec::<Controller>::new();
+        let ds4 = Controller::new();
         in_v.push(ds4);
-        DualShock4DriverManager {
+        ControllerManager {
             first_connector:ds4_conn, 
             connectors:ds4_conn_vec,
             controller_num:0, 
@@ -156,6 +157,11 @@ impl DualShock4DriverManager {
                 let s = i.clone();
                 dev_vec.push(s);
             }
+            else if i.vendor_id() == 1356 && i.product_id() == 3302
+            {
+                let s = i.clone();
+                dev_vec.push(s);
+            }
         }
 
         for d_i in dev_vec.clone()
@@ -174,22 +180,48 @@ impl DualShock4DriverManager {
             Some(dr)=>{
                 match dr.open_device(&self.api) {
                     Ok(device_)=>{
-                        let mut dsdr = DualShock4Driver{device:device_,mode:mode_, rgb:RGB::new()};
-                        dsdr.rgb.red = 0;
-                        dsdr.rgb.blue = 255;
-                        dsdr.rgb.grenn = 0;
+                        match dr.product_id()
+                        {
+                            2508=>{
+                                let mut controller = controller_driver::dualshock4::DualShock4Driver{device:device_, mode:mode_, rgb:RGB::new()};
 
-                        std::thread::spawn(move ||
-                            loop {
-                                let get = dsdr.task();
-                                
-                                let _ = publisher_.clone().send(get);
+                                controller.rgb = RGB::blue();
 
-                                if dsdr.mode == ControllerConnectionType::SERIAL
-                                {
-                                    dsdr.color_change()
-                                }
-                        });
+                                std::thread::spawn(move ||{
+                                    loop {
+                                        let get = controller.task();
+
+                                        let _ = publisher_.clone().send(get);
+
+                                        if controller.mode == ControllerConnectionType::SERIAL
+                                        {
+                                            controller.color_change()
+                                        }
+                                    }
+                                });
+                            }
+                            3302=>{
+                                let mut controller = controller_driver::dualsense::DualSenseDriver{device:device_, mode:ControllerConnectionType::BLE, rgb:RGB::new()};
+
+                                controller.rgb = RGB::blue();
+
+                                std::thread::spawn(move ||{
+                                    loop {
+                                        let get = controller.task();
+
+                                        let _ = publisher_.clone().send(get);
+
+                                        if controller.mode == ControllerConnectionType::SERIAL
+                                        {
+                                            controller.color_change()
+                                        }
+                                    }
+                                });
+                            }
+                            _=>{
+
+                            }
+                        }
                     }
                     Err(_e)=>{
 
@@ -202,39 +234,69 @@ impl DualShock4DriverManager {
         }
     }
 
-    pub fn add_driver(&mut self, mode_:ControllerConnectionType, publisher_:std::sync::mpsc::Sender<DualShock4>)
+    pub fn add_driver(&mut self, mode_:ControllerConnectionType, publisher_:std::sync::mpsc::Sender<Controller>)
     {
         match self.device_list.first()
         {
             Some(dr)=>{
                 match dr.open_device(&self.api) {
                     Ok(device_)=>{
-                        let mut dsdr = DualShock4Driver{device:device_,mode:mode_, rgb:RGB::new()};
-                        
-                        if !self.green_flag
+                        match dr.product_id()
                         {
-                            dsdr.rgb.grenn = 255;
-                            dsdr.rgb.blue = 0;
-                            dsdr.rgb.red = 0;
-                        }
-                        else {
-                            dsdr.rgb.grenn = 0;
-                            dsdr.rgb.blue = 0;
-                            dsdr.rgb.red = 255;
-                        }
+                            2508=>{
+                                let mut controller = controller_driver::dualshock4::DualShock4Driver{device:device_, mode:mode_, rgb:RGB::new()};
 
-                        std::thread::spawn(move ||
-                            loop {
-                                let get = dsdr.task();
-                                
-                                let _ = publisher_.clone().send(get);
-
-                                if dsdr.mode == ControllerConnectionType::SERIAL
+                                if self.green_flag
                                 {
-                                    dsdr.color_change();
+                                    controller.rgb = RGB::green()
                                 }
-                        });
+                                else 
+                                {
+                                    controller.rgb = RGB::red()
+                                }
 
+                                std::thread::spawn(move ||{
+                                    loop {
+                                        let get = controller.task();
+
+                                        let _ = publisher_.clone().send(get);
+
+                                        if controller.mode == ControllerConnectionType::SERIAL
+                                        {
+                                            controller.color_change()
+                                        }
+                                    }
+                                });
+                            }
+                            3302=>{
+                                let mut controller = controller_driver::dualsense::DualSenseDriver{device:device_, mode:ControllerConnectionType::BLE, rgb:RGB::new()};
+
+                                if self.green_flag
+                                {
+                                    controller.rgb = RGB::green()
+                                }
+                                else 
+                                {
+                                    controller.rgb = RGB::red()
+                                }
+
+                                std::thread::spawn(move ||{
+                                    loop {
+                                        let get = controller.task();
+
+                                        let _ = publisher_.clone().send(get);
+
+                                        if controller.mode == ControllerConnectionType::SERIAL
+                                        {
+                                            controller.color_change()
+                                        }
+                                    }
+                                });
+                            }
+                            _=>{
+
+                            }
+                        }
                     }
                     Err(_e)=>{
 
@@ -250,221 +312,8 @@ impl DualShock4DriverManager {
     }
 }
 
-pub struct DualShock4Driver
-{
-    device:HidDevice,
-    mode:ControllerConnectionType,
-    pub rgb:RGB
-}
 
-impl DualShock4Driver {
-    pub fn task(&mut self)->DualShock4
-    {
-            let mut buf = [0_u8;256];
-
-            match self.device.read(&mut buf) {
-                Ok(size)=>{
-                    let get_data = &buf[..size];
-                    let (j, btn, d) = convert(get_data, self.mode);
-
-                    if j.right_x == -0.9372549 && self.mode == ControllerConnectionType::BLE
-                    {
-                        self.mode = ControllerConnectionType::SERIAL
-                    }
-
-                    if j.left_x == 0.5058824 && self.mode == ControllerConnectionType::SERIAL
-                    {
-                        self.mode = ControllerConnectionType::BLE
-                    }
-
-                    DualShock4 {mode:self.mode, state:true, sticks: j, btns: btn, dpad: d }
-                }
-                Err(_)=>{
-                    DualShock4 {mode:self.mode,state:false, sticks:JoyStick::new(), btns:Buttons::new(), dpad:Dpad::new()}
-                }
-            }
-    }
-    pub fn color_change(&mut self)
-    {
-        if self.rgb.red == 0 && self.rgb.blue == 0 && self.rgb.grenn == 0
-        {
-            self.rgb.blue = 255;
-        }
-
-        let mut buf = [0u8; 32];
-        buf[0] = 0x05;
-        buf[1] = 0xFF;
-        buf[2] = 0x04;
-        buf[6] = self.rgb.red;
-        buf[7] = self.rgb.grenn;
-        buf[8] = self.rgb.blue;
-
-        match self.device.write(&buf) {
-            Ok(_d)=>{
-
-            }
-            Err(_e)=>{
-
-            }
-        }
-    }
-}
-
-fn convert(buf:&[u8], mode:ControllerConnectionType)->(JoyStick, Buttons, Dpad)
-{
-    if mode == ControllerConnectionType::BLE
-    {
-        let l_x = map(buf[3], 0.0, 255.0, -1.0, 1.0);
-        let l_y = map(buf[4], 0.0, 255.0, 1.0, -1.0);
-        let r_x = map(buf[5], 0.0, 255.0, -1.0, 1.0);
-        let r_y = map(buf[6], 0.0, 255.0, 1.0, -1.0);
-        let joy = JoyStick{left_x:l_x,left_y:l_y,right_x:r_x,right_y:r_y};
-        let mut btns = Buttons{
-            circle:false,
-            cross:false,
-            triangle:false,
-            cube:false,
-            r1:false,
-            r2:false,
-            l1:false,
-            l2:false,
-            left_push:false,
-            right_push:false,
-        };
-
-        let mut dpad = Dpad{
-            up_key:false,
-            down_key:false,
-            right_key:false,
-            left_key:false
-        };
-
-        match buf[7] {
-            0=>dpad.up_key = true,
-            2=>dpad.right_key = true,
-            4=>dpad.down_key = true,
-            6=>dpad.left_key = true,
-            24=>btns.cube = true,
-            40=>btns.cross = true,
-            72=>btns.circle = true,
-            136=>btns.triangle = true,
-            8=>(),
-            _=>()
-        }
-
-        match buf[8] {
-            1=>btns.l1 = true,
-            2=>btns.r1 = true,
-            4=>btns.l2 = true,
-            8=>btns.r2 = true,
-            64=>btns.left_push = true,
-            128=>btns.right_push = true,
-            _=>(),
-        }
-        (joy, btns, dpad)
-    }
-    else if mode == ControllerConnectionType::SERIAL
-    {
-        let l_x = map(buf[1], 0.0, 255.0, -1.0, 1.0);
-        let l_y = map(buf[2], 0.0, 255.0, 1.0, -1.0);
-        let r_x = map(buf[3], 0.0, 255.0, -1.0, 1.0);
-        let r_y = map(buf[4], 0.0, 255.0, 1.0, -1.0);
-        let joy = JoyStick{left_x:l_x,left_y:l_y,right_x:r_x,right_y:r_y};
-        let mut btns = Buttons{
-            circle:false,
-            cross:false,
-            triangle:false,
-            cube:false,
-            r1:false,
-            r2:false,
-            l1:false,
-            l2:false,
-            left_push:false,
-            right_push:false,
-        };
-
-        let mut dpad = Dpad{
-            up_key:false,
-            down_key:false,
-            right_key:false,
-            left_key:false
-        };
-
-        match buf[5] {
-            0=>dpad.up_key = true,
-            1=>{dpad.up_key = true;dpad.right_key = true},
-            2=>dpad.right_key = true,
-            3=>{dpad.right_key = true;dpad.down_key = true},
-            4=>dpad.down_key = true,
-            5=>{dpad.left_key=true;dpad.down_key=true},
-            6=>dpad.left_key = true,
-            7=>{dpad.left_key=true;dpad.up_key=true},
-            24=>btns.cube = true,
-            40=>btns.cross = true,
-            56=>{btns.cube=true;btns.cross=true},
-            72=>btns.circle = true,
-            88=>{btns.circle = true;btns.cube=true},
-            104=>{btns.circle=true;btns.cross=true},
-            136=>btns.triangle = true,
-            152=>{btns.cube=true;btns.triangle=true},
-            168=>{btns.triangle=true;btns.cross=true},
-            200=>{btns.triangle=true;btns.circle=true},
-            8=>(),
-            _=>()
-        }
-
-        match buf[6] {
-            1=>btns.l1 = true,
-            2=>btns.r1 = true,
-            4=>btns.l2 = true,
-            8=>btns.r2 = true,
-            64=>btns.left_push = true,
-            128=>btns.right_push = true,
-            192=>{btns.left_push = true; btns.right_push=true}
-            _=>(),
-        }
-        (joy, btns, dpad)
-    }
-    else {
-
-        let joy = JoyStick{left_x:0.0, left_y:0.0, right_x:0.0, right_y:0.0};
-        let btns = Buttons{
-            circle:false,
-            cross:false,
-            triangle:false,
-            cube:false,
-            r1:false,
-            r2:false,
-            l1:false,
-            l2:false,
-            left_push:false,
-            right_push:false,
-        };
-
-        let dpad = Dpad{
-            up_key:false,
-            down_key:false,
-            right_key:false,
-            left_key:false
-        };
-
-        (joy, btns, dpad)
-    }
-}   
-
-fn map(value:u8,in_min:f32, in_max:f32, out_min:f32, out_max:f32)->f32
-{
-    let mut result = (value as f32 - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
-
-    if result.abs() < 0.07
-    {
-        result = 0.0;
-    }
-
-    result
-}
-
-fn input_to_controller_view<'a>(input:DualShock4)->iced::widget::Row<'a,ControllerMessage>
+fn input_to_controller_view<'a>(input:Controller)->iced::widget::Row<'a,ControllerMessage>
 {
     let con_state = if input.state
             {
